pyproject.toml
[project]
name = "langchain-udemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.5",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]
